Add files via upload
After consulting with my professor, he consented to me posting the files that are not my original work, but are required to have the code function.

Motherofalltests provides examples of the proper way to call an instance of the function, in order to see the formatted output it produces. The same calls work for Calprint3 and Calprint4.